Major re-organization and split code by topic
